Re-impliment Home Button Functionality using atomic design principles. Closes #17
